### 2019-09-08 #1719 @whedon
[PRE REVIEW]: xrnet: Hierarchical Regularized Regression to Incorporate External Data
**Submitting author:** @gmweaver (<a href="http://orcid.org/0000-0002-9918-8386">Garrett Weaver</a>)
**Repository:** <a href="https://github.com/USCbiostats/xrnet" target ="_blank">https://github.com/USCbiostats/xrnet</a>
**Version:** 0.0.0.9000
