+++ paper/paper.bib
@@ -15,7 +15,8 @@
     volume = {33},
     number = {1},
     pages = {1--22},
-    url = {http://www.jstatsoft.org/v33/i01/}
+    url = {http://www.jstatsoft.org/v33/i01/},
+    doi = {10.18637/jss.v033.i01}
 }
 
 @Article{zeng2017,
@@ -36,7 +37,8 @@
     year = {2011},
     volume = {5},
     pages = {232--253},
-    number = {1}
+    number = {1},
+    doi = {doi:10.1214/10-AOAS388}
 }
 
 @Article{kane2013,
@@ -47,7 +49,8 @@
     volume = {55},
     number = {14},
     pages = {1--19},
-    url = {http://www.jstatsoft.org/v55/i14/}
+    url = {http://www.jstatsoft.org/v55/i14/},
+    doi = {10.18637/jss.v055.i14}
 }
 
 @MISC{eigenweb,
@@ -66,6 +69,7 @@
   number = {5},
   pages = {1--24},
   url = {http://www.jstatsoft.org/v52/i05/},
+  doi = {10.18637/jss.v052.i05}
 }
 
 @Article{rcpp,
