### 2019-09-11 65a107c
add cox solver
