+++ src/CoxSolver.h
@@ -1,0 +1,354 @@
+#ifndef COX_SOLVER_H
+#define COX_SOLVER_H
+
+#include <vector>
+#include "CoordSolver.h"
+
+using namespace Eigen;
+using namespace std;
+
+template <typename T>
+class CoxSolver : public CoordSolver<T> {
+
+    typedef Eigen::VectorXd VecXd;
+    typedef Eigen::VectorXi VecXi;
+    typedef Eigen::Map<const Eigen::MatrixXd> MapMat;
+    typedef Eigen::MappedSparseMatrix<double> MapSpMat;
+    typedef Eigen::Map<const Eigen::VectorXd> MapVec;
+
+private:
+    VecXd eta;
+    VecXd delta;
+    vector<double> D;
+    vector<double> d;
+    VecXi ri;
+    VecXi ck;
+    const int m;
+    using CoordSolver<T>::n;
+    using CoordSolver<T>::nv_total;
+    using CoordSolver<T>::intercept;
+    using CoordSolver<T>::wgts;
+    using CoordSolver<T>::wgts_user;
+    using CoordSolver<T>::wgts_sum;
+    using CoordSolver<T>::y;
+    using CoordSolver<T>::X;
+    using CoordSolver<T>::Fixed;
+    using CoordSolver<T>::XZ;
+    using CoordSolver<T>::xm;
+    using CoordSolver<T>::xs;
+    using CoordSolver<T>::xv;
+    using CoordSolver<T>::residuals;
+    using CoordSolver<T>::gradient;
+    using CoordSolver<T>::betas;
+    using CoordSolver<T>::b0;
+    using CoordSolver<T>::betas_prior;
+    using CoordSolver<T>::b0_prior;
+    using CoordSolver<T>::tolerance_irls;
+    using CoordSolver<T>::penalty;
+    using CoordSolver<T>::penalty_type;
+    using CoordSolver<T>::cmult;
+    using CoordSolver<T>::strong_set;
+
+
+public:
+    // constructor (dense X matrix)
+    CoxSolver(const Eigen::Ref<const Eigen::MatrixXd> & y_,
+              const Eigen::Ref<const Eigen::MatrixXd> & X_,
+              const Eigen::Ref<const Eigen::MatrixXd> & Fixed_,
+              const Eigen::Ref<const Eigen::MatrixXd> & XZ_,
+              const double * xmptr,
+              double * xvptr,
+              const double * xsptr,
+              VecXd wgts_user_,
+              bool intercept_,
+              const double * penalty_type_,
+              const double * cmult_,
+              VecXd quantiles_,
+              const double * ucl_,
+              const double * lcl_,
+              int ne_,
+              int nx_,
+              double tolerance_,
+              int max_iterations_) :
+              CoordSolver<T>(y_,
+                             X_,
+                             Fixed_,
+                             XZ_,
+                             xmptr,
+                             xvptr,
+                             xsptr,
+                             wgts_user_,
+                             intercept_,   //// want to set it false as default
+                             penalty_type_,
+                             cmult_,
+                             quantiles_,
+                             ucl_,
+                             lcl_,
+                             ne_,
+                             nx_,
+                             tolerance_,
+                             max_iterations_),
+                             eta(n),
+                             delta(n),
+                             ri(m+1),   //// this could be problem, m is unknown now
+                             ck(n+1)
+                             {
[…]
