### 2019-09-23 #1719 @jsgalan
Hi guys, I am currently out of office on a conference, I will be back by the end of the week, and can starting editing by next week.

Count me in!

Thanks
