### 2019-09-25 #1761 @juanvillada
Hi all!

This is a very useful piece of software accompanied by a very well written paper.

Some comments:

⚠️ When I tried to install the Master branch of `xrnet` on macOS (Mojave 10.14.6) I ran into the error:

```
ld: warning: directory not found for option '-L/usr/local/gfortran/lib'
```

✅ No problem with Master branch installation of `xrnet`  in Linux (Ubuntu 19.04)

✅ I am attaching the markdown file from the tests I did on the software. https://www.dropbox.com/s/j0x16cdmdzq7976/xrnet_run.html?dl=0

Thanks to the authors for open sourcing this tool! 👏🏼👏🏼
