### 2019-09-25 #1761 @gmweaver
Hi @juanvillada,

Thanks for the review!

With regard to installation on macOS, I believe this is specific to recent R versions that require a more recent version of GNU Fortran to compile packages from source. See https://cloud.r-project.org/bin/macosx/tools/ for a more detailed description and a binary to install the latest GNU Fortran, which should fix the installation issue. I can also update our documentation to inform users of this detail as well.

Many thanks again for your feedback!
