### 2019-09-26 #1761 @juanvillada
Hi @gmweaver,

Everything works fine after installing the latest GNU Fortran.

✅ No problem with Master branch installation of `xrnet` on macOS (Mojave 10.14.6).

Best wishes and thanks for this tool!
