### 2019-10-22 #1761 @jsgalan
Hi guys, this review escaped from email/alerts/calendars/todo-list. I am sorry.

The article explain the context very well and summarize the need for this tool and how it adds to the existing software packages.
- authors need to explain the acronyms SCAD and MCP or add some references that explains these methods.

Firstly I am using, R version 3.6.0 (2019-04-26) Platform: x86_64-apple-darwin15.6.0 (64-bit) on a MacOS Mojave 10.14.6

I tried the installation via master branch and got 17 warnings plus an installation error, see the resulting error
[xrnet.txt](https://github.com/openjournals/joss-reviews/files/3753851/xrnet.txt)

After installing the GNUfortran, as suggested by @gmweaver, I still got 17 warnings but a clean install, see the warnings here
[xrnet-gfortran.txt](https://github.com/openjournals/joss-reviews/files/3753883/xrnet-gfortran.txt)

It finally installed and I ran the demo script.

I have a few suggestions for the authors:

- add the GNUfortran installation part of the Setup in the front page of the package repository
- add a second and more complex example, for instance one reading a csv file.

Best,
