### 2019-10-23 #1761 @mgymrek
Thanks @jsgalan and @juanvillada for the reviews.

@gmweaver, can you post your response to these comments regarding the installation process here?
