### 2019-10-28 8fe2dc3
add citations to paper, update readme to clarify installation, add big.matrix example
