+++ paper/paper.bib
@@ -53,6 +53,28 @@
     doi = {10.18637/jss.v055.i14}
 }
 
+@Article{fan2001,
+    title = {Variable Selection via Nonconcave Penalized Likelihood and its Oracle Properties},
+    author = {Jianqing Fan and Runze Li},
+    journal = {Journal of the American Statistical Association},
+    year = {2001},
+    volume = {96},
+    number = {456},
+    pages = {1348--1360},
+    doi = {10.1198/016214501753382273}
+}
+
+@Article{zhang2010,
+    title = {Nearly unbiased variable selection under minimax concave penalty},
+    author = {Cun-Hui Zhang},
+    journal = {The Annals of Statistics},
+    year = {2010},
+    volume = {38},
+    number = {2},
+    pages = {894--942},
+    doi = {10.1214/09-AOS729}
+}
+
 @MISC{eigenweb,
   author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
   title = {Eigen v3},
