+++ inst/extdata/x_linear.txt
@@ -1,0 +1,200 @@
+0.0170688469099605,-1.63556168340907,-0.206736970939062,0.134563674310179,-0.96135049973924,0.8572621462461,0.865944703471274,0.518942626855688,0.0198291818172978,0.22466874592923,-0.255737081569969,-0.0473253906903541,-1.09350745691155,-0.260656212071231,-0.316057075867762,1.13752514106178,0.490197653940151,0.25126783346774,-0.0695816776430607,-1.39578591849092,1.55012914926562,3.26192430679691,1.79123733996135,0.28075669704919,-0.181451006654141,-0.714952138281742,-0.668840172727439,1.18419475566132,0.844742587776347,0.291897224995531,0.213077593864378,-1.15655891495833,-0.130249134465452,-0.26821964716182,-1.07631231537603,-1.56661951711695,-0.136961264382119,1.02979223687785,0.55252575966743,-0.159272564324847,-0.580752374928351,-0.333043313197982,0.667397361214023,1.47310099545455,0.190588291566973,-0.849788260530216,0.33148121324453,-0.0329142693558007,-0.194259127380137,-2.21356845689793
+-0.785211149106178,2.08333129517914,0.826046917067712,-1.46880652862918,-0.693106870276686,-1.39664937002973,1.18091746365826,0.572926912703744,-0.201712162805094,-0.954394933792221,0.0275653976859745,0.34542613271526,0.758261852275284,-0.67176058423815,0.325484108090393,-0.196540533826734,0.168951375556365,0.716641106931918,-0.00276640374084148,0.424912802915296,1.75351782961451,1.04467740376102,0.467533573532047,0.354443703472376,1.24509379647498,2.4780603305885,-0.269082982396595,0.390186500822384,-0.407186620945883,-0.895637677263892,-0.180738881446828,1.1866997255565,-0.0237559616446629,0.0835239041759103,-2.36690907279263,-0.718761660353604,0.541952344519457,-0.629425027136417,1.56344930600416,-0.590164250634317,-1.66322136109425,-0.332139329527845,-0.837561005874889,-0.96579829911895,0.939467536040897,0.877178019525138,0.151711565681842,0.645206115775801,-1.0580102528118,-1.75033718565427
+-0.51361551641116,0.0645204278218862,-1.88744451964002,-1.35970194851965,-0.388119268227991,0.351608504634896,0.076524640061823,0.872080803368565,0.154867479214046,-0.262713801640021,-0.689985989621464,-0.145366448154775,-0.0591096287925215,1.80680415262438,0.560138120417999,-2.10125196087668,-2.81550319968085,-1.17862733328994,1.21819989045072,0.70105173802541,1.3376256864747,1.17462818439894,1.25701496356984,0.531669269065487,1.52097793328463,-0.576003509923071,-1.78431251131189,1.18928091609108,-1.02100407919863,-0.37788494787483,0.383218250716634,-0.352951950256639,1.17730409528278,0.231957494266611,2.01419354662247,1.13096395688544,2.06937942559346,1.20917710468115,-0.590378300449202,-0.569062645929282,0.881241259791089,-0.502059948118542,-0.817516953117565,0.66496009872364,-0.422609931790792,-1.03835965774755,-0.40373890160068,-0.0530057527623472,1.56264497083512,-1.41551293198663
[…]

[…]
