### 2019-10-29 #1761 @gmweaver
@jsgalan thanks for the review and suggestions!

I have made the following changes based on your feedback:

- Added citations for the SCAD and MCP penalties to the paper
- Added more detail to the installation instructions regarding GNUFortran for Mac
- Added a second example that demonstrates how to use a big.matrix in the package by loading data from a text file. Let me know if this is enough of an addition.

The warnings when installing are specific to RcppEigen and although annoying, should be harmless. See https://stackoverflow.com/questions/49525561/rcppeigen-package-pragma-clang-diagnostic-pop-warnings for notes from one of the maintainers of that package (Dirk Eddelbeuttel)
