### 2019-10-29 #1761 @jsgalan
Hi @gmweaver,

- Please recompile the article via Whedon (whedon generate pdf) to see the changes.
- I checked the updated installation information, its way better.
- On the second example using big.matrix, is "x_linear.txt" a real file? could you provide an example of such file, even if its just 100,000 x 100,000 of random numbers?

Best,
