### 2019-10-29 60bed63
update bigmemory + xrnet example
