+++ src-i386/CoordSolver.h
@@ -1,0 +1,415 @@
+#ifndef COORD_SOLVER_H
+#define COORD_SOLVER_H
+
+#include <RcppEigen.h>
+#include "DataFunctions.h"
+#include <bigmemory/MatrixAccessor.hpp>
+// [[Rcpp::depends(RcppEigen, BH, bigmemory)]]
+
+template <typename T>
+class CoordSolver {
+
+    typedef Eigen::Map<const Eigen::MatrixXd> MapMat;
+    typedef Eigen::MappedSparseMatrix<double> MapSpMat;
+    typedef Eigen::Map<const Eigen::VectorXd> MapVec;
+    typedef Eigen::VectorXd VecXd;
+    typedef Eigen::VectorXi VecXi;
+
+protected:
+    const int n;
+    const int nv_total;
+    MapMat y;
+    double ym;
+    double ys;
+    T X;
+    MapMat Fixed;
+    MapMat XZ;
+    MapVec penalty_type;
+    MapVec cmult;
+    const VecXd quantiles;
+    MapVec ucl;
+    MapVec lcl;
+    const int ne;
+    const int nx;
+    const double tolerance;
+    const int max_iterations;
+    int num_passes;
+    double dlx;
+    VecXd penalty;
+    const bool intercept;
+    Eigen::Map<const Eigen::VectorXd> xm;
+    Eigen::Map<Eigen::VectorXd> xv;
+    Eigen::Map<const Eigen::VectorXd> xs;
+    VecXd wgts_user;
+    VecXd residuals;
+    VecXd wgts;
+    double wgts_sum;
+    VecXd betas;
+    VecXd betas_prior;
+    VecXd gradient;
+    double b0;
+    double b0_prior;
+    const double tolerance_irls;
+    Rcpp::LogicalVector strong_set;
+    Rcpp::LogicalVector active_set;
+    int status;
+    const double bigNum = 9.9e35;
+
+public:
+    // constructor (dense X matrix)
+    CoordSolver(const Eigen::Ref<const Eigen::MatrixXd> & y_,
+                const Eigen::Ref<const Eigen::MatrixXd> & X_,
+                const Eigen::Ref<const Eigen::MatrixXd> & Fixed_,
+                const Eigen::Ref<const Eigen::MatrixXd> & XZ_,
+                const double * xmptr,
+                double * xvptr,
+                const double * xsptr,
+                VecXd wgts_user_,
+                bool intercept_,
+                const double * penalty_type_,
+                const double * cmult_,
+                VecXd quantiles_,
+                const double * ucl_,
+                const double * lcl_,
+                int ne_,
+                int nx_,
+                double tolerance_,
+                int max_iterations_) :
+    n(X_.rows()),
+    nv_total(X_.cols() + Fixed_.cols() + XZ_.cols()),
+    y(y_.data(), n, y_.cols()),
+    ym(0.0),
+    ys(1.0),
+    X(X_.data(), n, X_.cols()),
+    Fixed(Fixed_.data(), n, Fixed_.cols()),
+    XZ(XZ_.data(), n, XZ_.cols()),
+    penalty_type(penalty_type_, nv_total),
+    cmult(cmult_, nv_total),
+    quantiles(quantiles_),
+    ucl(ucl_, nv_total),
+    lcl(lcl_, nv_total),
+    ne(ne_),
+    nx(nx_),
+    tolerance(tolerance_),
+    max_iterations(max_iterations_),
+    num_passes(0),
+    dlx(0.0),
+    penalty(2),
+    intercept(intercept_),
+    xm(xmptr, nv_total),
+    xv(xvptr, nv_total),
+    xs(xsptr, nv_total),
[…]
