+++ src-i386/BinomialSolver.h
@@ -1,0 +1,298 @@
+#ifndef BINOMIAL_SOLVER_H
+#define BINOMIAL_SOLVER_H
+
+#include "CoordSolver.h"
+
+template <typename T>
+class BinomialSolver : public CoordSolver<T> {
+
+    typedef Eigen::VectorXd VecXd;
+    typedef Eigen::VectorXi VecXi;
+    typedef Eigen::Map<const Eigen::MatrixXd> MapMat;
+    typedef Eigen::MappedSparseMatrix<double> MapSpMat;
+    typedef Eigen::Map<const Eigen::VectorXd> MapVec;
+
+private:
+    VecXd xbeta;
+    VecXd prob;
+    using CoordSolver<T>::n;
+    using CoordSolver<T>::nv_total;
+    using CoordSolver<T>::intercept;
+    using CoordSolver<T>::wgts;
+    using CoordSolver<T>::wgts_user;
+    using CoordSolver<T>::wgts_sum;
+    using CoordSolver<T>::y;
+    using CoordSolver<T>::X;
+    using CoordSolver<T>::Fixed;
+    using CoordSolver<T>::XZ;
+    using CoordSolver<T>::xm;
+    using CoordSolver<T>::xs;
+    using CoordSolver<T>::xv;
+    using CoordSolver<T>::residuals;
+    using CoordSolver<T>::gradient;
+    using CoordSolver<T>::betas;
+    using CoordSolver<T>::b0;
+    using CoordSolver<T>::betas_prior;
+    using CoordSolver<T>::b0_prior;
+    using CoordSolver<T>::tolerance_irls;
+    using CoordSolver<T>::penalty;
+    using CoordSolver<T>::penalty_type;
+    using CoordSolver<T>::cmult;
+    using CoordSolver<T>::strong_set;
+    const double prob_thresh = 1e-9;
+    double xbeta_thresh;
+
+
+public:
+    // constructor (dense X matrix)
+    BinomialSolver(const Eigen::Ref<const Eigen::MatrixXd> & y_,
+                   const Eigen::Ref<const Eigen::MatrixXd> & X_,
+                   const Eigen::Ref<const Eigen::MatrixXd> & Fixed_,
+                   const Eigen::Ref<const Eigen::MatrixXd> & XZ_,
+                   const double * xmptr,
+                   double * xvptr,
+                   const double * xsptr,
+                   VecXd wgts_user_,
+                   bool intercept_,
+                   const double * penalty_type_,
+                   const double * cmult_,
+                   VecXd quantiles_,
+                   const double * ucl_,
+                   const double * lcl_,
+                   int ne_,
+                   int nx_,
+                   double tolerance_,
+                   int max_iterations_) :
+    CoordSolver<T>(y_,
+                   X_,
+                   Fixed_,
+                   XZ_,
+                   xmptr,
+                   xvptr,
+                   xsptr,
+                   wgts_user_,
+                   intercept_,
+                   penalty_type_,
+                   cmult_,
+                   quantiles_,
+                   ucl_,
+                   lcl_,
+                   ne_,
+                   nx_,
+                   tolerance_,
+                   max_iterations_),
+                   xbeta(n),
+                   prob(n)
+                   {
+                       xbeta_thresh = log((1 - prob_thresh) / prob_thresh);
+                       init();
+                   };
+
+    // constructor (sparse X matrix)
+    BinomialSolver(const Eigen::Ref<const Eigen::MatrixXd> & y_,
+                   const MapSpMat X_,
+                   const Eigen::Ref<const Eigen::MatrixXd> & Fixed_,
[…]
