+++ src-i386/RcppExports.cpp
@@ -1,106 +1,0 @@
-// Generated by using Rcpp::compileAttributes() -> do not edit by hand
-// Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
-
-#include <RcppEigen.h>
-#include <Rcpp.h>
-
-using namespace Rcpp;
-
-// computeResponseRcpp
-Eigen::MatrixXd computeResponseRcpp(SEXP X, const int& mattype_x, const Eigen::Map<Eigen::MatrixXd> Fixed, const Eigen::Map<Eigen::VectorXd> beta0, const Eigen::Map<Eigen::MatrixXd> betas, const Eigen::Map<Eigen::MatrixXd> gammas, const std::string& response_type, const std::string& family);
-RcppExport SEXP _xrnet_computeResponseRcpp(SEXP XSEXP, SEXP mattype_xSEXP, SEXP FixedSEXP, SEXP beta0SEXP, SEXP betasSEXP, SEXP gammasSEXP, SEXP response_typeSEXP, SEXP familySEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< SEXP >::type X(XSEXP);
-    Rcpp::traits::input_parameter< const int& >::type mattype_x(mattype_xSEXP);
-    Rcpp::traits::input_parameter< const Eigen::Map<Eigen::MatrixXd> >::type Fixed(FixedSEXP);
-    Rcpp::traits::input_parameter< const Eigen::Map<Eigen::VectorXd> >::type beta0(beta0SEXP);
-    Rcpp::traits::input_parameter< const Eigen::Map<Eigen::MatrixXd> >::type betas(betasSEXP);
-    Rcpp::traits::input_parameter< const Eigen::Map<Eigen::MatrixXd> >::type gammas(gammasSEXP);
-    Rcpp::traits::input_parameter< const std::string& >::type response_type(response_typeSEXP);
-    Rcpp::traits::input_parameter< const std::string& >::type family(familySEXP);
-    rcpp_result_gen = Rcpp::wrap(computeResponseRcpp(X, mattype_x, Fixed, beta0, betas, gammas, response_type, family));
-    return rcpp_result_gen;
-END_RCPP
-}
-// fitModelCVRcpp
-Eigen::VectorXd fitModelCVRcpp(SEXP x, const int mattype_x, const Eigen::Map<Eigen::MatrixXd> y, SEXP ext, const bool& is_sparse_ext, const Eigen::Map<Eigen::MatrixXd> fixed, Eigen::VectorXd weights_user, const Rcpp::LogicalVector& intr, const Rcpp::LogicalVector& stnd, const Eigen::Map<Eigen::VectorXd> penalty_type, const Eigen::Map<Eigen::VectorXd> cmult, const Eigen::Map<Eigen::VectorXd> quantiles, const Rcpp::IntegerVector& num_penalty, const Rcpp::NumericVector& penalty_ratio, const Eigen::Map<Eigen::VectorXd> penalty_user, const Eigen::Map<Eigen::VectorXd> penalty_user_ext, Eigen::VectorXd lower_cl, Eigen::VectorXd upper_cl, const std::string& family, const std::string& user_loss, const Eigen::Map<Eigen::VectorXi> test_idx, const double& thresh, const int& maxit, const int& ne, const int& nx);
[…]
