+++ src-i386/rcpp_wrappers_cv.cpp
@@ -1,249 +1,0 @@
-#include <string.h>
-#include "DataFunctions.h"
-#include "XrnetCV.h"
-#include "XrnetUtils.h"
-#include "CoordDescTypes.h"
-#include "GaussianSolver.h"
-#include "BinomialSolver.h"
-
-template <typename TX, typename TZ>
-Eigen::VectorXd fitModelCV(const TX & x,
-                           const bool & is_sparse_x,
-                           const Eigen::Ref<const Eigen::MatrixXd> & y,
-                           const TZ & ext,
-                           const Eigen::Ref<const Eigen::MatrixXd> & fixed,
-                           Eigen::VectorXd weights_user,
-                           const Rcpp::LogicalVector & intr,
-                           const Rcpp::LogicalVector & stnd,
-                           const Eigen::Ref<const Eigen::VectorXd> & penalty_type,
-                           const Eigen::Ref<const Eigen::VectorXd> & cmult,
-                           const Eigen::Ref<const Eigen::VectorXd> & quantiles,
-                           const Rcpp::IntegerVector & num_penalty,
-                           const Rcpp::NumericVector & penalty_ratio,
-                           const Eigen::Ref<const Eigen::VectorXd> & penalty_user,
-                           const Eigen::Ref<const Eigen::VectorXd> & penalty_user_ext,
-                           Eigen::VectorXd lower_cl,
-                           Eigen::VectorXd upper_cl,
-                           const std::string & family,
-                           const std::string & user_loss,
-                           const Eigen::Ref<const Eigen::VectorXi> & test_idx,
-                           const double & thresh,
-                           const int & maxit,
-                           const int & ne,
-                           const int & nx) {
-
-    // initialize objects to hold means, variances, sds of all variables
-    const int n = x.rows();
-    const int nv_x = x.cols();
-    const int nv_fixed = fixed.size() == 0 ? 0 : fixed.cols();
-    const int nv_ext = ext.size() == 0 ? 0 : ext.cols();
-    const int nv_total = nv_x + nv_fixed + intr[1] + nv_ext;
-    Eigen::VectorXd xm = Eigen::VectorXd::Constant(nv_total, 0.0);
-    Eigen::VectorXd cent = Eigen::VectorXd::Constant(nv_total, 0.0);
-    Eigen::VectorXd xv = Eigen::VectorXd::Constant(nv_total, 1.0);
-    Eigen::VectorXd xs = Eigen::VectorXd::Constant(nv_total, 1.0);
-
-    // map to correct size of fixed matrix
-    Eigen::Map<const Eigen::MatrixXd> fixedmap(fixed.data(), fixed.rows(), nv_fixed);
-
-    // scale user weights
-    weights_user.array() = weights_user.array() / weights_user.sum();
-
-    // compute moments of matrices and create XZ (if external data present)
-    const bool center_x = intr[0] && !is_sparse_x;
-    compute_moments(x, weights_user, xm, cent, xv, xs, center_x, stnd[0], 0);
-    compute_moments(fixedmap, weights_user, xm, cent, xv, xs, center_x, stnd[0], nv_x);
-    const Eigen::MatrixXd xz = create_XZ(
-        x, ext, xm, cent, weights_user,
-        xv, xs, intr[1], stnd[1], nv_x + nv_fixed
-    );
-
-    // choose solver based on outcome
-    std::unique_ptr<CoordSolver<TX> > solver;
-    if (family == "gaussian") {
-        solver.reset(
-            new GaussianSolver<TX>(
-                    y, x, fixedmap, xz, cent.data(), xv.data(), xs.data(),
-                    weights_user, intr[0], penalty_type.data(),
-                    cmult.data(), quantiles, upper_cl.data(),
[…]
