### 2019-10-29 #1761 @gmweaver
@jsgalan "x_linear.txt" is a real file stored under inst/extdata in the repository (I followed the best practices suggested in the R Packages book by Hadley Wickham for including raw data files). I have updated the readme to explain where this file is located and and how to access within R after installing the package. Let me know if this will work or if you had something else in mind for a file example.
