### 2019-10-30 #1761 @jsgalan
hey @gmweaver that sounds great, thanks for the explanation. Also, the article checks great.

I am happy with the review.
