### 2019-11-01 #1761 @gmweaver
@jsgalan awesome, thanks for all the feedback!

@mgymrek I believe I have addressed both reviewers comments, let me know if there is anything else I can provide for the review
