### 2019-11-11 #1761 @mgymrek
Great. Thanks @jsgalan and @juanvillada for the reviews and @gmweaver for addressing all comments.

@gmweaver can you fix this minor typo in the paper? Everything else looks good:
"meta-features, also know as" -> "meta-features, also known as"

Then, please make a tagged release and archive (e.g. Zenodo) and report the version number and archive DOI here. Please ensure the title and author list match that in the paper.
