### 2019-11-11 940286c
fix paper type
