### 2019-11-11 #1761 @gmweaver
@mgymrek thanks for catching that typo! Fixed and generated the PDF again. I have also released a tagged version and archived to Zenodo.

The version is v0.1.1 and the Zenodo DOI is 10.5281/zenodo.3538245
