### 2019-11-11 080e54f
update version to match DESCRIPTION file
