### 2019-11-11 #1761 @gmweaver
@mgymrek small update to ensure DESCRIPTION file version matches tagged release version.

New version is v0.1.2 and Zenodo DOI is 10.5281/zenodo.3538278
