### 2019-11-12 #1761 @mgymrek
Thanks @gmweaver for the changes

@openjournals/joss-eics we are ready to accept this article!
