### 2019-12-05 #1761 @danielskatz
👋 @gmweaver - There are a few small issues with the paper, as indicated in https://github.com/USCbiostats/xrnet/pull/32.  Please merge this, or let me know which changes you disagree with.
