### 2019-12-05 #1761 @gmweaver
@danielskatz reviewed the changes in your pull request and everything looks good! Merged pull request and created a new release with changes.

New version is v0.1.3 and Zenodo DOI is 10.5281/zenodo.3564788
